+++ pkg/querysharding/analyzer.go
@@ -1,33 +1,38 @@
 // Copyright (c) The Thanos Authors.
 // Licensed under the Apache License 2.0.
 
+// Copyright 2013 The Prometheus Authors
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package querysharding
 
 import (
-	"fmt"
-
 	lru "github.com/hashicorp/golang-lru"
 	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/promql/parser"
 )
 
-// QueryAnalyzer is an analyzer which determines
-// whether a PromQL Query is shardable and using which labels.
-
 type Analyzer interface {
 	Analyze(string) (QueryAnalysis, error)
 }
 
+// QueryAnalyzer is an analyzer which determines
+// whether a PromQL Query is shardable and using which labels.
 type QueryAnalyzer struct{}
 
 type CachedQueryAnalyzer struct {
 	analyzer *QueryAnalyzer
 	cache    *lru.Cache
-}
-
-var nonShardableFuncs = []string{
-	"label_join",
-	"label_replace",
 }
 
 // NewQueryAnalyzer creates a new QueryAnalyzer.
@@ -80,14 +85,18 @@ func (a *QueryAnalyzer) Analyze(query string) (QueryAnalysis, error) {
 		return nonShardableQuery(), err
 	}
 
-	isShardable := true
-	var analysis QueryAnalysis
+	var (
+		analysis      QueryAnalysis
+		dynamicLabels []string
+	)
 	parser.Inspect(expr, func(node parser.Node, nodes []parser.Node) error {
 		switch n := node.(type) {
 		case *parser.Call:
-			if n.Func != nil && contains(n.Func.Name, nonShardableFuncs) {
-				isShardable = false
-				return fmt.Errorf("expressions with %s are not shardable", n.Func.Name)
+			if n.Func != nil {
+				if n.Func.Name == "label_join" || n.Func.Name == "label_replace" {
+					dstLabel := stringFromArg(n.Args[1])
+					dynamicLabels = append(dynamicLabels, dstLabel)
+				}
 			}
 		case *parser.BinaryExpr:
 			if n.VectorMatching != nil {
@@ -108,19 +117,42 @@ func (a *QueryAnalyzer) Analyze(query string) (QueryAnalysis, error) {
 		return nil
 	})
 
-	if !isShardable {
-		return nonShardableQuery(), nil
+	// If currently it is shard by, it is still shardable if there is
+	// any label left after removing the dynamic labels.
+	// If currently it is shard without, it is still shardable if we
+	// shard without the union of the labels.
+	// TODO(yeya24): we can still make dynamic labels shardable if we push
+	// down the label_replace and label_join computation to the store level.
+	if len(dynamicLabels) > 0 {
+		analysis = analysis.scopeToLabels(dynamicLabels, false)
 	}
 
 	return analysis, nil
 }
 
-func contains(needle string, haystack []string) bool {
-	for _, item := range haystack {
-		if needle == item {
-			return true
+// Copied from https://github.com/prometheus/prometheus/blob/v2.40.1/promql/functions.go#L1416.
+func stringFromArg(e parser.Expr) string {
+	tmp := unwrapStepInvariantExpr(e) // Unwrap StepInvariant
+	unwrapParenExpr(&tmp)             // Optionally unwrap ParenExpr
+	return tmp.(*parser.StringLiteral).Val
+}
+
+// Copied from https://github.com/prometheus/prometheus/blob/v2.40.1/promql/engine.go#L2642.
+// unwrapParenExpr does the AST equivalent of removing parentheses around a expression.
+func unwrapParenExpr(e *parser.Expr) {
+	for {
+		if p, ok := (*e).(*parser.ParenExpr); ok {
+			*e = p.Expr
+		} else {
+			break
 		}
 	}
+}
 
-	return false
+// Copied from https://github.com/prometheus/prometheus/blob/v2.40.1/promql/engine.go#L2652.
+func unwrapStepInvariantExpr(e parser.Expr) parser.Expr {
+	if p, ok := e.(*parser.StepInvariantExpr); ok {
+		return p.Expr
+	}
+	return e
 }